chore(trace): move sampler module
