+++ tests/contrib/aiohttp/test_middleware.py
@@ -10,7 +10,7 @@ from ddtrace.contrib.aiohttp.middlewares import CONFIG_KEY
 from ddtrace.contrib.aiohttp.middlewares import trace_app
 from ddtrace.contrib.aiohttp.middlewares import trace_middleware
 from ddtrace.ext import http
-from ddtrace.sampler import RateSampler
+from ddtrace.internal.sampler import RateSampler
 from tests.opentracer.utils import init_tracer
 from tests.utils import assert_span_http_status_code
 from tests.utils import flaky
